Refactor layout and add skeleton loaders
Reorganized layout components to separate filters and content panels. Introduced skeleton loaders for better user experience during content loading. Updated CSS to improve alignment and consistency across components.
