+++ app/projects/(projects-page)/@filters/loading.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { FiltersPanelSkeletonLoader } from '@/components/core/dashboard-pages-layout';
+
+const Loading = () => {
+  return (
+    <FiltersPanelSkeletonLoader />
+  );
+};
+
+export default Loading;
