+++ app/teams/(teams-page)/@content/page.module.css
@@ -1,21 +1,3 @@
-.team {
-  height: inherit;
-  display: flex;
-  position: relative;
-  background: #f1f5f9;
-  min-height: calc(100vh - 80px);
-}
-
-.team__left {
-  width: 0px;
-}
-
-.team__right {
-  position: relative;
-  flex: 1;
-  background: #f1f5f9;
-}
-
 .team__right__content {
   width: 100%;
   display: flex;
@@ -37,23 +19,10 @@
 }
 
 @media (min-width: 1024px) {
-  .team__right {
-    display: flex;
-    justify-content: center;
-    padding: 30px 0px 40px 0px;
-  }
-
   .team__right__teamslist {
     margin-top: 0px;
     min-height: calc(100vh - 325px);
     padding: 0px 16px;
-  }
-
-  .team__left {
-    width: 300px;
-    border-right: 1px solid #e2e8f0;
-    background: #fff;
-    height: calc(100vh - 80px);
   }
 
   .team__right__toolbar {
